+++ art-is-worship/src/app/services/contexted-core-content.service.ts
@@ -43,6 +43,7 @@ import { MumbaiMeriJaanList } from 'src/assets/gallery-files/lists-and-other/ima
 import { MumbaiMeriJaan2List } from 'src/assets/gallery-files/lists-and-other/image-lists/salaam-mumbai/mumbai-meri-jaan-2.list';
 import { ShadesAndShines1ImageList} from 'src/assets/gallery-files/lists-and-other/image-lists/special-lists/alag-grouping-full-cycle/shades-and-shines-1.image.list';
 import { ShadesAndShinesOtherColors1ImageList} from 'src/assets/gallery-files/lists-and-other/image-lists/special-lists/alag-grouping-full-cycle/shades-and-shines-other-colors-1.image.list';
+import { CharcoalTimesOtherColorsShinesImageList } from 'src/assets/gallery-files/lists-and-other/image-lists/special-lists/alag-grouping-full-cycle/charcoal-times-other-colors-shines.image.list';
 @Injectable({
   providedIn: 'root'
 })
@@ -116,7 +117,8 @@ export class ContextedCoreContentService {
   //  contentList.push( { contentFile:new SpecialAircraftSeriesImageList(), contentCategory:'majestic-lifters-legendary-carriers', roles:['non-living,  non-religious'],latest:true}) ;
     contentList.push( { contentFile:new MajesticLiftersLegendaryCarriersList(), contentCategory:'majestic-lifters-legendary-carriers', roles:['non-living,  non-religious'],reverseStack:'true',latest:true}) ;
     contentList.push( { contentFile:new ShadesAndShines1ImageList(),contentCategory:'shades-and-shines-1', roles:["any"],latest:true, reverseStack:'true'}) ;  
-    contentList.push( { contentFile:new ShadesAndShinesOtherColors1ImageList(),contentCategory:'shades-and-shines-other-colors-1', roles:["any"],latest:true, reverseStack:'true'}) ;          
+    contentList.push( { contentFile:new ShadesAndShinesOtherColors1ImageList(),contentCategory:'shades-and-shines-other-colors-1', roles:["any"],latest:true, reverseStack:'true'}) ;      
+    contentList.push( { contentFile:new CharcoalTimesOtherColorsShinesImageList(),contentCategory:'charcoal-times-other-colors-shines', roles:["any"],latest:true, reverseStack:'true'}) ;           
  }
   loadSpecialLists(contentList:ContentList[], specialFiles:any,strParam:string):boolean {