+++ art-is-worship/src/assets/gallery-files/lists-and-other/image-lists/special-lists/alag-grouping-full-cycle/charcoal-times-other-colors-shines.image.list.ts
@@ -1,0 +1,53 @@
+
+import {  ImageElement } from '@settings-and-models/image-detail'
+enum PicsIndex  {
+        aumkarGaneshSept2020Final,
+        AumGaShivParKashiVishwCTLightNov172020Final,
+        unsignedShivVeena12272020,
+        ShreeSiddhivinayakQ42020,
+        DeviMaaMay13182021,
+        MixGenPlaneSept2020,
+    };
+export class CharcoalTimesOtherColorsShinesImageList {
+  reverseStack =  'true';
+imageRoot = 'assets/gallery-files/images/special-lists/alag-grouping-full-cycle/charcoal-times-other-colors-shines/'; 
+public allImageList:ImageElement[]  =  [
+  {  
+            folder:'charcoal-times-other-colors-shines',
+            theme:'<theme>',
+            themeSummary: '',
+files: [
+{
+canvassMaterial:  'Plain paper',
+canvassSize:  'A4',
+content:  '',
+dateUploaded:  '03-04-2025',
+duplicate:  'false',
+summaryLabel:  `Colored Sangeet Ashtavinayaks 2 March 4 2025`,
+fullFileName: `${this.imageRoot}Sur-Niragas-Ho-Template-2-Mar4-2025.jpeg`,
+description: `<ul><li> Colored Sangeet Ashtavinayaks 2 March 4 2025 </li><li> 2nd of 8 Ganesh with musical instruments - Tanpura </li><li> Base template copied from B&W </li><li> Completion and upload March 4 2025 </li><li> Using color pencil, charcoal pencil, metal color pencil </li></ul>`,
+},
+{
+canvassMaterial:  'Plain paper',
+canvassSize:  'A4',
+content:  'color pencil, charcoal pencil, metal color pencil',
+dateUploaded:  '03-04-2025',
+duplicate:  'false',
+summaryLabel:  `Colored Sangeet Ashtavinayaks 1 March 4 2025`,
+fullFileName: `${this.imageRoot}Sur-Niragas-Ho-Template-1-Mar4-2025.jpeg`,
+description: `<ul><li> Colored Sangeet Ashtavinayaks 1 March 4 2025 </li><li> 1st of 8 Ganesh with musical instruments </li><li> Base template copied from B&W </li><li> Completion and upload March 4 2025 </li><li> Using color pencil, charcoal pencil, metal color pencil </li></ul>`,
+},
+{
+canvassMaterial:  'Plain paper',
+canvassSize:  'A4',
+content:  'black-white',
+dateUploaded:  '03-04-2025',
+duplicate:  'false',
+evolution:  `<b>First charcoal sketch <br/></B>`,
+evolutionSequence:  5,
+evolutionDate:  '03-04-2025',
+summaryLabel:  `Ganesh using Charcoal pencils March 3 2025`,
+fullFileName: `${this.imageRoot}Ganesh-Charcoal-March-3-2025.jpeg`,
+description: `<ul><li> Ganesh using Charcoal pencils March 3 2025 </li><li> Using Kalour Hard, Medium and Soft black, and white charcoal pencils </li><li> These pencils were gifted by my wife long back and I finally got down to using them  </li><li> Uploaded March 4 2025 </li></ul>`,
+},
+  ]}]}